Move from GPL2 to Apache2 license
The Apache2 license used by many library dependencies in not compatible with the GPL2

As such the project is moving to use the Apache2 license for its own works, thus
correcting the issue and making the project compatible with the licensing of its libraries.

I am the sole author of the entire codebase up to this point and give permission for the
project to migrate licenses. I do not own, nor do I claim to own the rights to the software
dependencies included except where they are given to me.